WebAssembly: HACK: disable generating lldb-optimized debug info
WebAssembly doesn't have support for the DWARF5 features yet.

It's probably not too hard to add (just add the sections), but I
decided to just disable it out of caution since I have no idea
if other changes are needed.

This doesn't test for the WebAssembly platform yet; will do that later