[stdlib] Fix return type of swift_{uint64,int64,float*}ToString
The return type of these functions are uint64_t in Stubs.cpp
but UInt in the Swift code; this changes the Swift code to match
the C++ return type.

Found when compiling the stdlib for WebAssembly, which requires that
all return types match: UInt maps to i32 while uint64_t maps to i64,
so functions calling these functions fail the validation.